% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAvoidableEDAtt.R
\name{calculateAvoidableEDAtt}
\alias{calculateAvoidableEDAtt}
\title{Derive avoidable ED attendances as per Uni of Sheffield ScHARR avoidable attendances metric}
\usage{
calculateAvoidableEDAtt(
  typeED,
  attDisposal,
  attCategory,
  arrivalMode,
  attInvestigations,
  attTreatments,
  clinicalCodeStandard = "snomed",
  numThreads = 1
)
}
\arguments{
\item{typeED}{list/vector: type of ED attendance}

\item{attDisposal}{list/vector: attendance disposal}

\item{attCategory}{list/vector: attendance category}

\item{arrivalMode}{list/vector: arrival mode}

\item{attInvestigations}{data.frame/data.table: all ED investigations}

\item{attTreatments}{data.frame/data.table: all ED treatments}

\item{clinicalCodeStandard}{character; either snomed or hes}

\item{numThreads}{integer; number of threads to use}
}
\value{
data.frame with single column containing binary TRUE/FALSE whether avoidable attendance
}
\description{
This returns a data.frame with logical vector (TRUE/FALSE) corresponding to
each of the input rows, stating whether the (patient level) attendance is
'avoidable' as per the University of Sheffield's School of Health and Related
Research (ScHARR)'s avoidable attendance metric. \link{insert reference here}.
The code will derive this for either SNOMED or HES codes, and is based upon
codes published by NHS Digital.
}
\examples{
calculateAvoidableEDAtt(df[,1],df[,2],df[,3],df[,4],df[,5:6],df[,7:9],'snomed',3)
calculateAvoidableEDAtt(dt[,.(edType)],dt[,.(attDisp)],dt[,.(attCat)],dt[,.(attMode)],dt[,.(inves1,inves2)],dt[,.(treat1,treat2,treat3)],'hes',1)
}
