% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setESAAvoidableAttThreads}
\alias{setESAAvoidableAttThreads}
\title{Set the number of threads to use by default}
\usage{
setESAAvoidableAttThreads(n)
}
\arguments{
\item{n}{integer of number of threads to use}
}
\description{
Set the number of threads to use by default
}
