% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAvoidableEDAtt.R
\name{calculateAvoidableEDAtt}
\alias{calculateAvoidableEDAtt}
\title{Derive avoidable ED attendances as per Uni of Sheffield ScHARR avoidable attendances metric}
\usage{
calculateAvoidableEDAtt(
  data,
  colTypeED,
  colAttDischarge,
  colAttCategory,
  colArrivalMode,
  colsInvestigations,
  colsTreatments,
  clinicalCodeStandard = "snomed",
  nthreads = getOption("ESAAvoidableAtt_nthreads")
)
}
\arguments{
\item{data}{data.frame/data.table: attendance-level data.frame containing ED data}

\item{colTypeED}{string: column name of the type of ED attendance}

\item{colAttDischarge}{string: column name of the attendance discharge/disposal}

\item{colAttCategory}{string: column name of the  attendance category}

\item{colArrivalMode}{string: column name of the arrival mode}

\item{colsInvestigations}{character vector: column names of all the ED investigations}

\item{colsTreatments}{character vector: column names of all the ED treatments}

\item{clinicalCodeStandard}{string; clinical coding standard to use - either snomed or hes}

\item{nthreads}{integer; number of threads to use}
}
\value{
logical vector containing binary TRUE/FALSE whether each row is an avoidable attendance
}
\description{
This returns a data.frame with logical vector (TRUE/FALSE) corresponding to
each of the input rows, stating whether the (patient level) attendance is
'avoidable' as per the University of Sheffield's School of Health and Related
Research (ScHARR)'s avoidable attendance metric. \link{insert reference here}.
The code will derive this for either SNOMED or HES codes, and is based upon
codes published by NHS Digital.
}
\examples{
calculateAvoidableEDAtt(dt,'Department_Type','Discharge_Status','AttendanceCategory','Arrival_Mode',paste0('Investigation_',1:30),paste0('Treatment_',1:30),'snomed',30)
calculateAvoidableEDAtt(dt,'Department_Type','Discharge_Status','AttendanceCategory','Arrival_Mode',paste0('Investigation_',1:30),paste0('Treatment_',1:30),'hes',6)
}
